refactor: fold intellihide into dock auto hide service
