+++ DockDoor/Utilities/DockUtils.swift
@@ -56,29 +56,235 @@ class DockUtils {
 }
 
 final class DockAutoHideManager {
+    static let shared = DockAutoHideManager()
+
+    private let pollInterval: TimeInterval = 0.15
+    private let transitionHideDuration: TimeInterval = 0.75
+    private let titleBarHideDuration: TimeInterval = 1.1
+    private let minimumFrameDeltaToDetectTransition: CGFloat = 2
+    private let titleBarDetectionHeight: CGFloat = 80
+
     private var wasAutoHideEnabled: Bool?
-    private var isManagingDock: Bool = false
+    private var isManagingDock = false
+    private var previewRequestsVisibleDock = false
+    private var lastAppliedAutoHideState: Bool?
+
+    private var defaultsObserver: Defaults.Observation?
+    private var evaluationTimer: Timer?
+    private var hasPendingRefresh = false
+    private let geometry = DockIntellihideGeometry()
+    private lazy var titleBarClickMonitor = DockTitleBarClickMonitor { [weak self] location in
+        self?.handleTitleBarDoubleClick(at: location)
+    }
+
+    private var lastSample: DockIntellihideWindowSample?
+    private var forceHideUntil: Date?
+
+    private init() {
+        defaultsObserver = Defaults.observe(keys: .enableDockIntellihide) { [weak self] in
+            DispatchQueue.main.async {
+                self?.reconfigureIntellihide()
+            }
+        }
+        reconfigureIntellihide()
+    }
+
+    deinit {
+        defaultsObserver = nil
+        stopEvaluating()
+        titleBarClickMonitor.stop()
+        restoreManagedDockState(force: true)
+    }
 
     func preventDockHiding(_ windowSwitcherActive: Bool = false) {
-        guard Defaults[.preventDockHide], !windowSwitcherActive else { return }
-        let currentAutoHideState = CoreDockGetAutoHideEnabled()
-
-        if currentAutoHideState {
-            wasAutoHideEnabled = currentAutoHideState
-            isManagingDock = true
-            CoreDockSetAutoHideEnabled(false)
-        }
+        previewRequestsVisibleDock = Defaults[.preventDockHide] && !windowSwitcherActive
+        refreshNow()
     }
 
     func restoreDockState() {
-        if isManagingDock, let wasEnabled = wasAutoHideEnabled {
-            CoreDockSetAutoHideEnabled(wasEnabled)
-            wasAutoHideEnabled = nil
-            isManagingDock = false
+        previewRequestsVisibleDock = false
+        refreshNow()
+    }
+
+    func refreshNow() {
+        guard Defaults[.enableDockIntellihide] || previewRequestsVisibleDock || isManagingDock else { return }
+
+        if Thread.isMainThread {
+            applyCurrentPolicy()
+            return
+        }
+
+        guard !hasPendingRefresh else { return }
+        hasPendingRefresh = true
+        DispatchQueue.main.async { [weak self] in
+            guard let self else { return }
+            hasPendingRefresh = false
+            applyCurrentPolicy()
         }
     }
 
     func cleanup() {
-        restoreDockState()
+        previewRequestsVisibleDock = false
+        stopEvaluating()
+        titleBarClickMonitor.stop()
+        lastSample = nil
+        forceHideUntil = nil
+        restoreManagedDockState(force: true)
+    }
+
+    private func reconfigureIntellihide() {
+        if Defaults[.enableDockIntellihide] {
+            startEvaluating()
+            titleBarClickMonitor.start()
+        } else {
+            stopEvaluating()
+            titleBarClickMonitor.stop()
+            lastSample = nil
+            forceHideUntil = nil
+        }
+
+        refreshNow()
+    }
+
+    private func startEvaluating() {
+        guard evaluationTimer == nil else { return }
+
+        let timer = Timer(timeInterval: pollInterval, repeats: true) { [weak self] _ in
+            self?.applyCurrentPolicy()
+        }
+        evaluationTimer = timer
+        RunLoop.main.add(timer, forMode: .common)
+    }
+
+    private func stopEvaluating() {
+        evaluationTimer?.invalidate()
+        evaluationTimer = nil
+    }
+
+    private func applyCurrentPolicy() {
+        if previewRequestsVisibleDock {
+            applyAutoHide(false)
+            return
+        }
+
+        guard Defaults[.enableDockIntellihide] else {
+            restoreManagedDockState()
+            return
+        }
+
+        evaluateDockVisibility()
+    }
+
+    private func captureOriginalDockStateIfNeeded() {
+        if !isManagingDock {
+            wasAutoHideEnabled = CoreDockGetAutoHideEnabled()
+            isManagingDock = true
+        }
+    }
+
+    private func restoreManagedDockState(force: Bool = false) {
+        guard force || !previewRequestsVisibleDock, isManagingDock, let wasAutoHideEnabled else { return }
+        CoreDockSetAutoHideEnabled(wasAutoHideEnabled)
+        self.wasAutoHideEnabled = nil
+        isManagingDock = false
+        lastAppliedAutoHideState = nil
+    }
+
+    private func evaluateDockVisibility() {
+        guard let frontmostApp = NSWorkspace.shared.frontmostApplication,
+              frontmostApp.activationPolicy == .regular,
+              frontmostApp.bundleIdentifier != Bundle.main.bundleIdentifier,
+              frontmostApp.bundleIdentifier != "com.apple.dock"
+        else {
+            lastSample = nil
+            forceHideUntil = nil
+            applyAutoHide(false)
+            return
+        }
+
+        guard let sample = geometry.sample(for: frontmostApp) else {
+            if WindowUtil.isAppInFullscreen(frontmostApp) {
+                applyAutoHide(true)
+                return
+            }
+            lastSample = nil
+            forceHideUntil = nil
+            applyAutoHide(false)
+            return
+        }
+
+        if sample.isFullscreen {
+            applyAutoHide(true)
+            return
+        }
+
+        guard let windowScreen = geometry.screen(for: sample.frame),
+              let dockContext = geometry.dockContext(preferCached: lastAppliedAutoHideState == true || forceHideUntil != nil)
+        else {
+            applyAutoHide(false)
+            return
+        }
+
+        guard windowScreen.uniqueIdentifier() == dockContext.screen.uniqueIdentifier() else {
+            forceHideUntil = nil
+            applyAutoHide(false)
+            return
+        }
+
+        if geometry.shouldForceHideTransition(
+            from: lastSample,
+            to: sample,
+            dockInfluenceRegion: dockContext.releaseRegion,
+            minimumFrameDeltaToDetectTransition: minimumFrameDeltaToDetectTransition
+        ) {
+            forceHideUntil = Date().addingTimeInterval(transitionHideDuration)
+        }
+
+        lastSample = sample
+
+        if let forceHideUntil, forceHideUntil > Date() {
+            applyAutoHide(true)
+            return
+        }
+        forceHideUntil = nil
+
+        let shouldHide = if lastAppliedAutoHideState == true {
+            sample.frame.intersects(dockContext.releaseRegion)
+        } else {
+            sample.frame.intersects(dockContext.dockRegion)
+        }
+
+        applyAutoHide(shouldHide)
+    }
+
+    private func handleTitleBarDoubleClick(at location: CGPoint) {
+        guard Defaults[.enableDockIntellihide],
+              let frontmostApp = NSWorkspace.shared.frontmostApplication,
+              frontmostApp.activationPolicy == .regular,
+              frontmostApp.bundleIdentifier != Bundle.main.bundleIdentifier,
+              let sample = geometry.sample(for: frontmostApp)
+        else {
+            return
+        }
+
+        if geometry.titleBarRegion(for: sample.frame, detectionHeight: titleBarDetectionHeight).contains(location) {
+            forceHideUntil = Date().addingTimeInterval(titleBarHideDuration)
+            applyAutoHide(true)
+        }
+    }
+
+    private func applyAutoHide(_ enabled: Bool, force: Bool = false) {
+        captureOriginalDockStateIfNeeded()
+
+        if !force, lastAppliedAutoHideState == enabled {
+            return
+        }
+
+        let currentState = CoreDockGetAutoHideEnabled()
+        if force || currentState != enabled {
+            CoreDockSetAutoHideEnabled(enabled)
+        }
+
+        lastAppliedAutoHideState = enabled
     }
 }
